Revert "test: move ci visibility integration tests"
This reverts commit 37d3f98db8321668b43f1511368fa39d71bc3e14.
