+++ tests/integration/test_integration.py
@@ -16,7 +16,12 @@ from ddtrace.constants import AUTO_KEEP
 from ddtrace.constants import SAMPLING_PRIORITY_KEY
 from ddtrace.internal import agent
 from ddtrace.internal import compat
+from ddtrace.internal.ci_visibility.constants import AGENTLESS_ENDPOINT
 from ddtrace.internal.ci_visibility.constants import COVERAGE_TAG_NAME
+from ddtrace.internal.ci_visibility.constants import EVP_PROXY_AGENT_ENDPOINT
+from ddtrace.internal.ci_visibility.constants import EVP_SUBDOMAIN_HEADER_EVENT_VALUE
+from ddtrace.internal.ci_visibility.constants import EVP_SUBDOMAIN_HEADER_NAME
+from ddtrace.internal.ci_visibility.recorder import CIVisibility
 from ddtrace.internal.ci_visibility.writer import CIVisibilityWriter
 from ddtrace.internal.encoding import JSONEncoder
 from ddtrace.internal.encoding import MsgpackEncoderV03 as Encoder
@@ -527,6 +532,38 @@ def test_priority_sampling_rate_honored(encoding, monkeypatch):
         ), "the proportion of sampled spans should approximate the sample rate given by the agent"
 
         t.shutdown()
+
+
+@pytest.mark.skipif(AGENT_VERSION == "testagent", reason="Test agent doesn't support evp proxy.")
+def test_civisibility_intake_with_evp_available():
+    with override_env(dict(DD_API_KEY="foobar.baz", DD_SITE="foo.bar")):
+        with override_global_config({"_ci_visibility_agentless_enabled": False}):
+            t = Tracer()
+            CIVisibility.enable(tracer=t)
+            assert CIVisibility._instance.tracer._writer._endpoint == EVP_PROXY_AGENT_ENDPOINT
+            assert CIVisibility._instance.tracer._writer.intake_url == agent.get_trace_url()
+            assert (
+                CIVisibility._instance.tracer._writer._headers[EVP_SUBDOMAIN_HEADER_NAME]
+                == EVP_SUBDOMAIN_HEADER_EVENT_VALUE
+            )
+            CIVisibility.disable()
+
+
+def test_civisibility_intake_with_missing_apikey():
+    with override_env(dict(DD_SITE="foobar.baz")):
+        with override_global_config({"_ci_visibility_agentless_enabled": True}):
+            with pytest.raises(EnvironmentError):
+                CIVisibility.enable()
+
+
+def test_civisibility_intake_with_apikey():
+    with override_env(dict(DD_API_KEY="foobar.baz", DD_SITE="foo.bar")):
+        with override_global_config({"_ci_visibility_agentless_enabled": True}):
+            t = Tracer()
+            CIVisibility.enable(tracer=t)
+            assert CIVisibility._instance.tracer._writer._endpoint == AGENTLESS_ENDPOINT
+            assert CIVisibility._instance.tracer._writer.intake_url == "https://citestcycle-intake.foo.bar"
+            CIVisibility.disable()
 
 
 def test_bad_endpoint():
